chore: Refactor component imports and file structure
