+++ src/components/common/auto-log-route/AutoLogRoute.types.ts
