+++ src/components/common/auto-log-route/index.ts
