+++ src/components/common/copyright/Copyright.types.ts
