+++ src/components/common/copyright/index.ts
