+++ src/components/common/index.ts
@@ -1,0 +1,3 @@
+export { default as Copyright } from './copyright';
+export { default as PrivateRoute } from './private-route';
+export { default as AutoLogRoute } from './auto-log-route';
