--- src/components/index.ts
@@ -1,5 +1,0 @@
-export * from './copyright';
-export * from './signin';
-export * from './sign-up';
-export * from './private-route';
-export { default as AutoLogRoute } from './auto-log-route';
