+++ src/components/pages/sign-up/SignUp.types.ts
