+++ src/context/index.ts
@@ -1,1 +1,1 @@
-export { default } from './auth';
+export { default as AuthProvider } from './auth';
